# docs/openapi/ironwall-api.yaml
openapi: "3.1.0"
info:
  title: AXELUS-WAF API
  version: "1.0.0"
  description: |
    # AXELUS-WAF REST API

    Complete API reference for the AXELUS-WAF platform by **OPTIMIUM NEXUS LLC**.

    ## Authentication

    Most endpoints require one of:
    - **Bearer Token** (JWT): `Authorization: Bearer <token>`
    - **API Key**: `X-Admin-Key: <key>` (admin endpoints) or `X-API-Key: <key>` (portal endpoints)

    ## Base URLs

    | Environment | URL |
    |------------|-----|
    | Management API | `https://YOUR_WAF_HOST:9443` |
    | License Server | `http://YOUR_WAF_HOST:8090` |
    | Customer Portal | `http://YOUR_WAF_HOST:8080` |
    | Admin Center | `http://YOUR_WAF_HOST:8085` |

    ## Rate Limits

    All API endpoints are rate-limited. Responses include:
    - `X-RateLimit-Limit`: Request limit per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

  contact:
    name: OPTIMIUM NEXUS LLC
    url: https://www.optimiumnexus.com
    email: contact@optimiumnexus.com
  license:
    name: Proprietary
    url: https://www.optimiumnexus.com/license
servers:
  - url: https://{host}:9443
    description: Management API
    variables:
      host:
        default: localhost
        description: Your AXELUS server hostname

  - url: http://{host}:8090
    description: License Server API
    variables:
      host:
        default: localhost

  - url: http://{host}:8080
    description: Customer Portal API
...
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin API key for license and admin endpoints

    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login

    PortalAPIKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Customer portal API key

  # ── Schemas ───────────────────────────────────────────────────────────────────
  schemas:
...
    AnomalyScore:
      type: object
      properties:
        score:               { type: number, minimum: 0, maximum: 1 }
        level:               { type: string, enum: [normal, suspect, anomaly, zero_day] }
        is_zero_day:         { type: boolean }
        confidence:          { type: number }
        baseline_deviation:  { type: number }
        explanation:         { type: string }
        top_features:
          type: array
          items:
            type: object
            properties:
              feature:   { type: string }
              value:     { type: number }
              deviation: { type: number }
              weight:    { type: number }
...
paths:
...
  /api/open/zerodayshield/score:
    post:
      tags: [Zero-Day Shield]
      summary: Score a request for anomaly detection
      description: Runs the request through the Isolation Forest ML model and returns an anomaly score.
      operationId: scoreRequest
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ip:         { type: string }
                method:     { type: string }
                path:       { type: string }
                query:      { type: string }
                body:       { type: string }
                user_agent: { type: string }
                headers:
                  type: object
                  additionalProperties: { type: string }
      responses:
        '200':
          description: Anomaly score result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomalyScore'
...